Move interactive source tree behavior to its own class
